Better movement
Migrated to new input system and made movement more abrupt

+++ SpaceInvadersGame/Assets/_Presentation/Player/PlayerMovement.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class PlayerMovement
 {
-    public void MovePlayer(Transform player, float horizontalDirection, float speed)
+    public void MovePlayer(Transform player, int horizontalDirection, float speed)
     {
         player.position += new Vector3(horizontalDirection * speed * Time.deltaTime, 0f, 0f);
     }
